pyproject.toml
[project]
name = "daedalus"
version = "0.5.0"
description = "Architecture of Incarnation — Distributed Autonomous Evolving Dense Architecture for Living Unified Self"
authors = [{name = "Massimo Azzano"}]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "unsloth>=2024.8",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.5.0",
    "FlagEmbedding>=1.2.0",
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "hdbscan>=0.8.33",
    "numpy>=1.26.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "datasets>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "night*", "eval*", "scripts*"]
